case_study/backend/regulatory/api.py
"""
API endpoints for compliance and regulatory requirements.

Implement the following endpoints:
- GET /facilities/{facility_id}/compliance - Get compliance status for a facility
- POST /compliance/calculate - Calculate compliance obligation

Note: Use the sample_data from the main api.py file directly. No database operations are required.
"""
from ninja import Router
from django.http import JsonResponse
from typing import List, Dict, Any

# Create a router for this app
router = Router()

# Access the sample data from the main api.py file
# from compliance_projection.api import sample_data

# TODO: Implement the following endpoints:
# 1. GET /facilities/{facility_id}/compliance - Get compliance status for a facility
# 2. POST /compliance/calculate - Calculate compliance obligation
