case_study/backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
contourpy==1.3.2
cycler==0.12.1
Django==5.2
django-cors-headers==4.7.0
django-ninja==1.4.1
fonttools==4.57.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
